Made PartController and PartService handle exceptions

+++ src/main/java/com/example/examinbackend/service/PartService.java
@@ -5,6 +5,7 @@ import com.example.examinbackend.repository.PartRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PartService {
@@ -15,8 +16,13 @@ public class PartService {
         this.partRepository = partRepository;
     }
 
-    public Part getPartById(Long id) {
-        return partRepository.findById(id).orElse(null);
+    public Optional<Part> getPartById(Long id) {
+        Optional<Part> optionalPart = partRepository.findById(id);
+        if (optionalPart.isEmpty()) {
+            return Optional.empty();
+        }else {
+            return optionalPart;
+        }
     }
     public Part getPartByName(String name) {
         return partRepository.findByPartName(name).orElse(null);
@@ -25,14 +31,22 @@ public class PartService {
         return partRepository.save(part);
     }
 
-    public void deletePart(Long id) {
+    public Optional<Part> deletePart(Long id) {
+        Optional<Part> optionalPart = getPartById(id);
+        if (optionalPart.isEmpty()) {
+            return Optional.empty();
+        }
         partRepository.deleteById(id);
+        return optionalPart;
     }
 
-    public Part updatePartName(Long id, Part part) {
-        Part existingPart = getPartById(id);
-        existingPart.setPartName(part.getPartName());
-        return partRepository.save(existingPart);
+    public Optional<Part> updatePartName(Long id, Part part) {
+        Optional<Part> existingPart = getPartById(id);
+        if (existingPart.isEmpty()) {
+            return Optional.empty();
+        }
+        existingPart.get().setPartName(part.getPartName());
+        return Optional.of(partRepository.save(existingPart.get()));
     }
 
     public List<Part> getAllParts() {
